refactor(private-router): split barrel into per-domain routers
Break the 1432-line src/index.ts barrel into domain folders under
src/routers/ (connection, queueSource, overview, queues, jobs, flows,
workers, schedulers). Routers stay thin and delegate to service.ts
functions; types and schemas live next to their handlers. Shared
foundations (tRPC init, source contract, queue-target, guards) move to
src/trpc.ts, src/source.ts, and src/shared/.

index.ts is now a thin aggregator that re-exports the identical public
surface, so the unchanged test suite (18/18) and all consumers
(embedded-core, frontend, standalone) typecheck and pass unchanged.
