+++ packages/private-router/src/routers/flows/types.ts
@@ -1,0 +1,16 @@
+export type FlowListInput =
+  | {
+      queueKey?: string;
+      queueName?: string;
+      prefix?: string;
+      limit?: number;
+    }
+  | undefined;
+
+export type FlowTargetInput = {
+  queueKey?: string;
+  queueName?: string;
+  name?: string;
+  prefix?: string;
+  flowId: string;
+};
