+++ packages/private-router/src/routers/queueSource/router.ts
@@ -1,0 +1,8 @@
+import type { PrivateDashboardQueueSource } from "../../source";
+import { authenticatedProcedure, t } from "../../trpc";
+
+export function createQueueSourceRouter(source: PrivateDashboardQueueSource) {
+  return t.router({
+    status: authenticatedProcedure.query(() => source.getStatus()),
+  });
+}
