+++ packages/private-router/src/routers/workers/types.ts
@@ -1,0 +1,14 @@
+export type WorkerListInput = {
+  queueKey?: string;
+  queueName?: string;
+  prefix?: string;
+  limit?: number;
+};
+
+export type WorkerTargetInput = {
+  queueKey?: string;
+  queueName?: string;
+  name?: string;
+  prefix?: string;
+  workerId: string;
+};
